refactor, start mapi package and move config struct from driver implementation
